WorkignOnSalesOrder

+++ app.BusinessLogic/SalesOrderDetailServices/SalesOrderDetailViewModel.cs
@@ -14,7 +14,7 @@ namespace app.Services.SalesOrderDetailServices
         public long UnitId { get; set; }
         public Unit Unit { get; set; }
         public string UnitName { get; set; }
-        public decimal SalesQty { get; set; }
+        public int SalesQty { get; set; }
         public decimal SalesPrice { get; set; }
         public decimal Discount { get; set; }
         public decimal TotalAmount { get; set; }
